+++ src/app/core/help/help-routing.module.ts
@@ -4,7 +4,7 @@ import { ActivatedRouteSnapshot, Resolve, RouterModule } from '@angular/router';
 import { HelpTopics } from './help-topic.component';
 import { HelpComponent } from './help.component';
 import { HelpTopicWrapperComponent } from './help-topic-wrapper.component';
-import { AuthGuard } from '../../core/auth/auth.guard';
+import { AuthGuard } from '../auth/auth.guard';
 
 @Injectable()
 export class HelpBreadcrumbResolver implements Resolve<string> {
