+++ src/app/core/help/help.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute, Event, NavigationEnd, Router } from '@angular/router';
 
 import { filter } from 'rxjs/operators';
 
-import { BreadcrumbService, Breadcrumb } from '../breadcrumb/breadcrumb.service';
+import { BreadcrumbService, Breadcrumb } from '../../common/breadcrumb/breadcrumb.service';
 import { HelpTopics } from './help-topic.component';
 
 export class HelpTopic {
