+++ src/app/site/example/help/example-help.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { HelpTopics } from '../../../common/help/help-topic.component';
+import { HelpTopics } from '../../../core/help/help-topic.component';
 
 @Component({
 	template: '<p>Example help content.</p>'
